+++ NEWS.md
@@ -1,5 +1,7 @@
 # pw (development version)
 
+* `pw_test()` stops if the playwright skeleton is missing
+
 * Functions to test if npx is in the PATH
 
 * Initial version with `pw_init()` & `pw_test()`